Adds async option for finding path size.
Well, this is disappointing. The async version has worse performance
across the board, according to the benchmarks. It may be that our
benchmarks do not capture situations where it might shine
(e.g. our dir branch factor of 2 is too low), or whatever benefits
exist do not overcome the thread overhead.

In any case, support is added for now for benchmarking, since it
is possible async becomes valuable after other potential changes
(e.g. performing the sorting on-the-fly.) But it will likely be
removed eventually if its value cannot be demonstrated.