Remove Ord instance
Turns out, the added Ord instance has really poor performance.
The new sorting logic (size the+n path) is still slower than the
original (sorting by size only), but it is not nearly as slow as the
Ord instance (using all fields). The reason we do not use the new
(size + path) is that it would be unlawful e.g. we could have
x <= y && y <= x && x /= y.