Disable non-moving gc in benchmarks
It turns out, the non-moving gc + -with-rtsopts=-N + the asyncPool
benchmark was causing the massive performance regressions (even in the
non-asyncPool ones!). Presumably the culprit is memory black-magic.
Disabling the non-moving gc fixes this.

Incidentally, we only enabled it because tasty-bench recommended it,
but that recommendation has since been rescinded.